+++ src/services/BlockRepository.ts
@@ -3,7 +3,9 @@ import {injectable} from 'inversify';
 import {HexStringWith0x} from '../types/custom';
 import {v4 as uuid} from 'uuid';
 import Block from '../models/Block';
-import Leaf from '../models/Leaf';
+import Leaf from '../types/Leaf';
+import {BigNumber} from 'ethers';
+import {SignedBlockConsensus} from '../types/Consensus';
 
 type Params = {
   id?: string;
@@ -17,7 +19,20 @@ type Params = {
 };
 
 @injectable()
-class SaveMintedBlock {
+class BlockRepository {
+  async saveBlock(chainAddress: string, consensus: SignedBlockConsensus, blockId: BigNumber): Promise<void> {
+    await this.apply({
+      id: `block::${blockId}`,
+      chainAddress,
+      dataTimestamp: new Date(consensus.dataTimestamp * 1000),
+      timestamp: new Date(),
+      leaves: consensus.leaves,
+      blockId: blockId.toNumber(),
+      root: consensus.root,
+      fcdKeys: consensus.fcdKeys,
+    });
+  }
+
   async apply(params: Params): Promise<Block> {
     const block = new Block();
     block._id = params.id || uuid();
@@ -28,25 +43,12 @@ class SaveMintedBlock {
     block.root = params.root;
     block.data = this.treeDataFor(params.leaves);
     block.fcdKeys = params.fcdKeys;
-    await this.attachLeavesToBlock(params.leaves, params.blockId);
-    return await getModelForClass(Block).create(block);
-  }
 
-  private async attachLeavesToBlock(leaves: Leaf[], blockId: number): Promise<void> {
-    return getModelForClass(Leaf)
-      .updateMany(
-        {
-          _id: {
-            $in: leaves.map((leaf) => leaf._id),
-          },
-        },
-        {
-          $set: {
-            blockId: blockId,
-          },
-        },
-      )
-      .exec();
+    return getModelForClass(Block).findOneAndUpdate({blockId: block.blockId}, block, {
+      upsert: true,
+      setDefaultsOnInsert: true,
+      new: true,
+    });
   }
 
   private treeDataFor(leaves: Leaf[]): Record<string, HexStringWith0x> {
@@ -54,4 +56,4 @@ class SaveMintedBlock {
   }
 }
 
-export default SaveMintedBlock;
+export default BlockRepository;