+++ src/services/SortedMerkleTreeFactory.ts
@@ -1,13 +1,13 @@
 import {injectable} from 'inversify';
 
-import Leaf from '../models/Leaf';
+import Leaf from '../types/Leaf';
 import {KeyValuePairs} from '../types/custom';
 import {SortedMerkleTree} from '@umb-network/toolbox';
 import {remove0x} from '@umb-network/toolbox/dist/utils/helpers';
 
 @injectable()
 class SortedMerkleTreeFactory {
-  apply(sortedLeaves: Leaf[]): SortedMerkleTree {
+  static apply(sortedLeaves: Leaf[]): SortedMerkleTree {
     const treeData: KeyValuePairs = {};
 
     sortedLeaves.forEach((leaf) => {