+++ src/types/Consensus.ts
@@ -1,5 +1,6 @@
-import Leaf from '../models/Leaf';
+import Leaf from '../types/Leaf';
 import {BigNumber} from 'ethers';
+import Feeds from './Feed';
 
 export interface DataForConsensus {
   affidavit: string;
@@ -18,4 +19,28 @@ export interface Consensus {
   fcdValues: number[];
   signatures: string[];
   power: BigNumber;
+}
+
+export interface SignedBlockConsensus {
+  dataTimestamp: number;
+  leaves: Leaf[];
+  root: string;
+  fcdKeys: string[];
+}
+
+export interface ProposedConsensus {
+  dataTimestamp: number;
+  fcds: Leaf[];
+  root: string;
+  signer: string;
+  fcdKeys: string[];
+  leaves: Leaf[];
+  affidavit: string;
+}
+
+export interface LeavesAndFeeds {
+  firstClassLeaves: Leaf[];
+  leaves: Leaf[];
+  fcdsFeeds: Feeds;
+  leavesFeeds: Feeds;
 }