+++ src/types/Leaf.ts
@@ -1,0 +1,6 @@
+import {HexStringWith0x} from './custom';
+
+export default interface Leaf {
+  label: string;
+  valueBytes: HexStringWith0x;
+}