+++ test/services/BlockRepository.test.ts
@@ -5,13 +5,15 @@ import mongoose from 'mongoose';
 import * as uuid from 'uuid';
 
 import {loadTestEnv} from '../helpers/loadTestEnv';
-import SaveMintedBlock from '../../src/services/SaveMintedBlock';
+import BlockRepository from '../../src/services/BlockRepository';
 import Block from '../../src/models/Block';
-import Leaf from '../../src/models/Leaf';
+import Leaf from '../../src/types/Leaf';
 import {getModelForClass} from '@typegoose/typegoose';
+import {BigNumber} from 'ethers';
+import {SignedBlockConsensus} from '../../src/types/Consensus';
 
-describe('SaveMintedBlock', () => {
-  let saveMintedBlock: SaveMintedBlock;
+describe('BlockRepository', () => {
+  let blockRepository: BlockRepository;
 
   before(async () => {
     const config = loadTestEnv();
@@ -20,23 +22,21 @@ describe('SaveMintedBlock', () => {
 
   beforeEach(async () => {
     await getModelForClass(Block).deleteMany({});
-    await getModelForClass(Leaf).deleteMany({});
-    saveMintedBlock = new SaveMintedBlock();
+    blockRepository = new BlockRepository();
   });
 
   after(async () => {
     await getModelForClass(Block).deleteMany({});
-    await getModelForClass(Leaf).deleteMany({});
     await mongoose.connection.close();
   });
 
   it("builds tree data and saves in block's object", async () => {
     const leaves: Leaf[] = [
-      {_id: uuid.v4(), label: 'ETH-USD', valueBytes: '0x01', timestamp: new Date(), blockId: 1},
-      {_id: uuid.v4(), label: 'USD-ETH', valueBytes: '0x02', timestamp: new Date(), blockId: 1},
+      {label: 'ETH-USD', valueBytes: '0x01'},
+      {label: 'USD-ETH', valueBytes: '0x02'},
     ];
 
-    const result = await saveMintedBlock.apply({
+    const result = await blockRepository.apply({
       id: 'block::1',
       chainAddress: '0x333',
       blockId: 1,
@@ -54,7 +54,7 @@ describe('SaveMintedBlock', () => {
   });
 
   it('generates UUID and saves the block to database', async () => {
-    await saveMintedBlock.apply({
+    await blockRepository.apply({
       chainAddress: '0x333',
       blockId: 1,
       leaves: [],
@@ -73,5 +73,21 @@ describe('SaveMintedBlock', () => {
     expect(blockFromDb?.data).to.satisfies((data: any) => data === undefined || Object.keys(data).length === 0);
     expect(blockFromDb?.root).to.be.eq('0x00');
     expect(blockFromDb?.timestamp).to.be.a('Date');
+  });
+
+  describe('#saveBlock', () => {
+    it('saves a block with its leaves', async () => {
+      const blockConsensus: SignedBlockConsensus = {
+        dataTimestamp: 1,
+        leaves: [],
+        root: '0x00',
+        fcdKeys: ['ETH-USD', 'USD-ETH'],
+      };
+
+      await blockRepository.saveBlock('0x333', blockConsensus, BigNumber.from(1));
+
+      const blocksCount = await getModelForClass(Block).countDocuments({}).exec();
+      expect(blocksCount).to.be.eq(1);
+    });
   });
 });